Update integration tests to .net core 3
