+++ tests/Microsoft.Identity.Test.Integration/HeadlessTests/CancellationTests.cs
