+++ tests/Microsoft.Identity.Test.Integration/HeadlessTests/OboTests.cs
