+++ tests/Microsoft.Identity.Test.Integration/HeadlessTests/RegionalAuthIntegrationTests.cs
