+++ tests/Microsoft.Identity.Test.Integration/HeadlessTests/SilentAuthTests.cs
