+++ tests/Microsoft.Identity.Test.Integration/Infrastructure/ClientAppTestExtensions.cs
@@ -3,6 +3,7 @@
 
 using System.Diagnostics;
 using Microsoft.Identity.Client;
+using Microsoft.Identity.Test.Common;
 
 namespace Microsoft.Identity.Test.Integration.net45.Infrastructure
 {
