+++ tests/Microsoft.Identity.Test.Integration/Infrastructure/HttpTelemetryRecorder.cs
