+++ tests/Microsoft.Identity.Test.Integration/Infrastructure/MsalAssert.cs
