+++ tests/Microsoft.Identity.Test.Integration/SeleniumTests/ConfidentialClientAuthorizationTests.cs
