+++ tests/Microsoft.Identity.Test.Integration/SeleniumTests/FociTests.cs
