+++ tests/Microsoft.Identity.Test.Integration/SeleniumTests/SSHCertificateTests.cs
