+++ tests/Microsoft.Identity.Test.Integration/SeleniumTests/SeleniumInfrastructureTests.cs
